vm/requirements.txt
pyyaml>=6.0
requests>=2.31.0
MetaTrader5>=5.0.45